+++ fltk/src/app/state.rs
@@ -28,6 +28,8 @@ impl<T: Sync + Send + 'static> GlobalState<T> {
     }
 
     /// Modifies the global state by acquiring a mutable reference
+    /// # Panics
+    /// Panics on state downcast errors
     pub fn with<V: Clone, F: 'static + Fn(&mut T) -> V>(&self, cb: F) -> V {
         if let Some(val) = STATE.get().unwrap().lock().unwrap().downcast_mut::<T>() {
             cb(val)