+++ fltk/src/macros/window.rs
@@ -11,7 +11,7 @@ macro_rules! impl_window_ext {
                     #[cfg(feature = "rwh05")]
                     type Handle = Win32WindowHandle;
                     #[cfg(feature = "raw-window-handle")]
-                    type Handle = Win32Handle::empty();
+                    type Handle = Win32Handle;
                     let mut handle = Handle::empty();
                     handle.hwnd = self.raw_handle();
                     handle.hinstance = $crate::app::display();
@@ -75,6 +75,59 @@ macro_rules! impl_window_ext {
                         let mut handle = Handle::empty();
                         handle.surface = self.raw_handle();
                         return RawWindowHandle::Wayland(handle);
+                    }
+                }
+            }
+        }
+
+        #[cfg(feature = "rwh05")]
+        unsafe impl HasRawDisplayHandle for $name {
+            fn raw_display_handle(&self) -> RawDisplayHandle {
+                #[cfg(target_os = "windows")]
+                {
+                    type Handle = WindowsDisplayHandle;
+                    let handle = Handle::empty();
+                    return RawDisplayHandle::Windows(handle);
+                }
+
+                #[cfg(target_os = "macos")]
+                {
+                    type Handle = AppKitDisplayHandle;
+                    let handle = Handle::empty();
+                    return RawDisplayHandle::AppKit(handle);
+                }
+
+                #[cfg(target_os = "android")]
+                {
+                    type Handle = AndroidDisplayHandle;
+                    let handle = Handle::empty();
+                    return RawDisplayHandle::Android(handle);
+                }
+
+                #[cfg(any(
+                    target_os = "linux",
+                    target_os = "dragonfly",
+                    target_os = "freebsd",
+                    target_os = "netbsd",
+                    target_os = "openbsd",
+                ))]
+                {
+                    #[cfg(not(feature = "use-wayland"))]
+                    {
+                        type Handle = XlibDisplayHandle;
+                        let mut handle = Handle::empty();
+                        handle.display = $crate::app::display();
+                        handle.screen = self.screen_num();
+                        return RawDisplayHandle::Xlib(handle);
+                    }
+
+
+                    #[cfg(feature = "use-wayland")]
+                    {
+                        type Handle = WaylandDisplayHandle;
+                        let mut handle = Handle::empty();
+                        handle.display = $crate::app::display();
+                        return RawDisplayHandle::Wayland(handle);
                     }
                 }
             }