+++ fltk/src/text.rs
@@ -764,7 +764,7 @@ pub enum TextAttr {
 }
 
 /// Defines the styles used in the `set_highlight_data`, which is used with style buffers
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct StyleTableEntry {
     /// Font color
     pub color: Color,
@@ -774,8 +774,18 @@ pub struct StyleTableEntry {
     pub size: i32,
 }
 
+impl Default for StyleTableEntry {
+    fn default() -> Self {
+        Self {
+            color: Color::Foreground,
+            font: Font::Helvetica,
+            size: crate::app::font_size(),
+        }
+    }
+}
+
 /// Defines the styles used in the `set_highlight_data`, which is used with style buffers
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub struct StyleTableEntryExt {
     /// Font color
     pub color: Color,
@@ -787,6 +797,18 @@ pub struct StyleTableEntryExt {
     pub attr: TextAttr,
     /// background color
     pub bgcolor: Color,
+}
+
+impl Default for StyleTableEntryExt {
+    fn default() -> Self {
+        Self {
+            color: Color::Foreground,
+            font: Font::Helvetica,
+            size: crate::app::font_size(),
+            attr: TextAttr::None,
+            bgcolor: Color::Background2,
+        }
+    }
 }
 
 impl TextEditor {